requirements.txt
flask-mysql==1.4.0